fix: split updateData into the mutations file
