+++ packages/frontend/components/StockForm.tsx
@@ -9,33 +9,8 @@ type StockFormProps = {
 
 const StockForm = (props: StockFormProps) => {
   const { newStock, setNewStock, paints } = React.useContext(GlobalContext);
-  const { updatePaintMutation } = usePaintMutations();
-  const updateData = async (result: any, paint: Paint, paints: Paint[]) => {
-    if (result === undefined) {
-      updatePaintMutation.mutate({
-        id: paint.id,
-        name: paint.name,
-        status: paint.status,
-        updatedAt: new Date(),
-        stock: paint.stock,
-      });
-    } else {
-      updatePaintMutation.mutate({
-        id: result.draggableId,
-        status: result.destination.droppableId,
-        updatedAt: new Date(),
-        stock: paint.stock,
-      });
-      const items = [...paints];
+  const { updatePaintMutation, updateData } = usePaintMutations();
 
-      const [reorderedItem] = items.splice(result.source.index, 1);
-
-      items.splice(result.destination.index, 0, reorderedItem);
-
-      const idsOrderArray = items.map((task) => task.id);
-      localStorage.setItem("paintOrder", JSON.stringify(idsOrderArray));
-    }
-  };
   return (
     <dialog id="StockModal" className="modal">
       <form method="dialog" className="modal-box">
@@ -57,7 +32,12 @@ const StockForm = (props: StockFormProps) => {
         <div className="modal-action">
           <button
             onClick={() => {
-              updateData(props.destination, { stock: newStock }, paints);
+              updateData(
+                props.destination,
+                { stock: newStock },
+                paints,
+                updatePaintMutation
+              );
             }}
             className="btn"
           >
