+++ packages/frontend/mutations/paint.tsx
@@ -1,4 +1,5 @@
 import { updatePaint, createPaint, deletePaint } from "@/lib/paintApi";
+import { Paint } from "@/typings";
 import { useMutation, useQueryClient } from "react-query";
 import { toast } from "react-toastify";
 
@@ -36,5 +37,43 @@ export const usePaintMutations = () => {
       toast.error("Error deleting paint", e.response.data.message);
     },
   });
-  return { updatePaintMutation, createPaintMutation, deletePaintMutation };
+
+  const updateData = async (
+    result: any,
+    paint: Paint,
+    paints: Paint[],
+    updatePaintMutation: any
+  ) => {
+    if (result === undefined) {
+      updatePaintMutation.mutate({
+        id: paint.id,
+        name: paint.name,
+        status: paint.status,
+        updatedAt: new Date(),
+        stock: paint.stock,
+      });
+    } else {
+      updatePaintMutation.mutate({
+        id: result.draggableId,
+        status: result.destination.droppableId,
+        updatedAt: new Date(),
+        stock: paint.stock,
+      });
+      const items = [...paints];
+
+      const [reorderedItem] = items.splice(result.source.index, 1);
+
+      items.splice(result.destination.index, 0, reorderedItem);
+
+      const idsOrderArray = items.map((task) => task.id);
+      localStorage.setItem("paintOrder", JSON.stringify(idsOrderArray));
+    }
+  };
+
+  return {
+    updatePaintMutation,
+    createPaintMutation,
+    deletePaintMutation,
+    updateData,
+  };
 };
